backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1
pydantic==2.5.0
pydantic-settings==2.1.0
python-jose[cryptography]==3.3.0
passlib==1.7.4
bcrypt==4.1.2
python-multipart==0.0.6
authlib==1.2.1
httpx==0.25.2
python-dotenv==1.0.0
langchain==0.1.0
langchain-community==0.0.10
ollama==0.1.6
faiss-cpu==1.7.4
pypdf==3.17.1
python-docx==1.1.0
tiktoken==0.5.1
numpy==1.26.2
scikit-learn==1.3.2
pandas==2.1.3
pydantic[email]==2.5.0
